+++ app/pb_kits/playbook/pb_date_picker/docs/_date_picker_default.jsx
@@ -1,40 +1,12 @@
-import React, { useState } from 'react'
-import { Date as DateKit, DatePicker, LabelValue } from '../../'
+import React from 'react'
+import { DatePicker } from '../../'
 
-const DatePickerDefault = () => {
-  const [dateString, setDateString] = useState((new Date()).toLocaleDateString())
-  const [dateObj, setDateObj] = useState([new Date()])
+const DatePickerDefault = () => (
+  <div>
+    <DatePicker
+        pickerId="date-picker-default"
+    />
+  </div>
+)
 
-  const changeHandler = (dateStr, selectedDates) => {
-    setDateString(dateStr)
-    setDateObj(selectedDates)
-  }
-
-  return (
-    <div>
-      <LabelValue
-          label="Date Object"
-          value={dateObj[0] ? dateObj[0].toString() : ''}
-      />
-      <br />
-      <br />
-      <LabelValue
-          label="Date String"
-          value={dateString}
-      />
-      <br />
-      <br />
-      <DateKit
-          value={dateObj[0]}
-      />
-      <br />
-      <br />
-      <DatePicker
-          defaultDate={dateString}
-          onChange={changeHandler}
-          pickerId="date-picker-default"
-      />
-    </div>
-  )
-}
 export default DatePickerDefault
